[fs] xfs: avoid CIL allocation during insert
Message-id: <1379494347-28456-94-git-send-email-dchinner@redhat.com>
Patchwork-id: 67385
O-Subject: [RHEL 7.0 PATCH 093/155] xfs: avoid CIL allocation during insert
Bugzilla: 1001861
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

RH BZ: 1001861
Upstream commit f5baac354db8b6abfe8ed4ff6b6c3438c42ea606
--

From: Dave Chinner <dchinner@redhat.com>

Now that we have the size of the log vector that has been allocated,
we can determine if we need to allocate a new log vector for
formatting and insertion. We only need to allocate a new vector if
it won't fit into the existing buffer.

However, we need to hold the CIL context lock while we do this so
that we can't race with a push draining the currently queued log
vectors. It is safe to do this as long as we do GFP_NOFS allocation
to avoid avoid memory allocation recursing into the filesystem.
Hence we can safely overwrite the existing log vector on the CIL if
it is large enough to hold all the dirty regions of the current
item.

Signed-off-by: Dave Chinner <dchinner@redhat.com>
Reviewed-by: Mark Tinguely <tinguely@sgi.com>
Signed-off-by: Ben Myers <bpm@sgi.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>