[fs] xfs: btree block LSN escaping to disk uninitialised
Message-id: <1379494347-28456-135-git-send-email-dchinner@redhat.com>
Patchwork-id: 67461
O-Subject: [RHEL 7.0 PATCH 134/155] xfs: btree block LSN escaping to disk uninitialised
Bugzilla: 1001861
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

RH BZ: 1001861
Upstream commit b58fa554e9b940083a0691f7234c13240fc09377
--

From: Dave Chinner <dchinner@redhat.com>

When testing LSN ordering code for v5 superblocks, it was discovered
that the the LSN embedded in the generic btree blocks was
occasionally uninitialised. These values didn't get written to disk
by metadata writeback - they got written by previous transactions in
log recovery.

The issue is here that the when the block is first allocated and
initialised, the LSN field was not initialised - it gets overwritten
before IO is issued on the buffer - but the value that is logged by
transactions that modify the header before it is written to disk
(and initialised) contain garbage. Hence the first recovery of the
buffer will stamp garbage into the LSN field, and that can cause
subsequent transactions to not replay correctly.

The fix is simply to initialise the bb_lsn field to zero when we
initialise the block for the first time.

Signed-off-by: Dave Chinner <dchinner@redhat.com>
Reviewed-by: Mark Tinguely <tinguely@sgi.com>
Signed-off-by: Ben Myers <bpm@sgi.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>