[drm] qxl: add delayed fb operations
Message-id: <1379642638-5668-1-git-send-email-airlied@redhat.com>
Patchwork-id: 67549
O-Subject: [PATCH] [RHEL7.0 PATCH] drm/qxl: add delayed fb operations
Bugzilla: 1002056
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1002056
upstream: 0665f9f852b4ac05f2f62046a169f0f5b9212317

Due to the nature of qxl hw we cannot queue operations while in an irq
context, so we queue these operations as best we can until atomic allocations
fail, and dequeue them later in a work queue.

Daniel looked over the locking on the list and agrees it should be sufficent.

The atomic allocs use no warn, as the last thing we want if we haven't memory
to allocate space for a printk in an irq context is more printks.

Signed-off-by: Dave Airlie <airlied@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>