[tracing] Add function probe to trigger a ftrace dump of current CPU trace
Message-id: <1380033213-24659-3-git-send-email-jolsa@redhat.com>
Patchwork-id: 67713
O-Subject: [PATCH RHEL7.0 BZ1011527 2/2] tracing: Add function probe to trigger a ftrace dump of current CPU trace
Bugzilla: 1011527
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1011527
https://bugzilla.redhat.com/show_bug.cgi?id=1011527

upstream status
===============
commit 90e3c03c3a09a7b176b3fe59d78f5d9755ac8e37
Author: Steven Rostedt (Red Hat) <rostedt@goodmis.org>
Date:   Tue Apr 30 19:00:46 2013 -0400

description
===========
Add the "cpudump" command to have the current CPU ftrace buffer dumped
to console if a function is hit. This is useful when debugging a
tripple fault, where you have an idea of a function that is called
just before the tripple fault occurs, and can tell ftrace to dump its
content out to the console before it continues.

This differs from the "dump" command as it only dumps the content of
the ring buffer for the currently executing CPU, and does not show
the contents of the other CPUs.

Format is:

  <function>:cpudump

echo 'bad_address:cpudump' > /debug/tracing/set_ftrace_filter

To remove this:

echo '!bad_address:cpudump' > /debug/tracing/set_ftrace_filter

Signed-off-by: Jarod Wilson <jarod@redhat.com>