[tools] perf/diff: Switching the base hists to be pairs head
Message-id: <1380034318-25505-5-git-send-email-jolsa@redhat.com>
Patchwork-id: 67723
O-Subject: [PATCH RHEL7.0 BZ1011529 04/12] perf diff: Switching the base hists to be pairs head
Bugzilla: 1011529
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1011529
https://bugzilla.redhat.com/show_bug.cgi?id=1011529

upstream status
===============
commit 9af303e22a317d1cc6f440e08f72428830708b37
Author: Jiri Olsa <jolsa@redhat.com>
Date:   Sat Dec 1 21:15:40 2012 +0100

description
===========
Making the baseline hists to act as a pairs head.

So far we don't care which hists act as a pairs head, because we have
only 2 files to deal with and any of them is suitable to do the job.

But if we want to process more files, we need to pick up one hists to
act as pairs head, and the baseline hists is the most suitable.

Signed-off-by: Jarod Wilson <jarod@redhat.com>