[security] keys: Consolidate the concept of an 'index key' for key access
Message-id: <20131002135314.13612.32397.stgit@warthog.procyon.org.uk>
Patchwork-id: 68294
O-Subject: [RHEL7 PATCH 04/33] KEYS: Consolidate the concept of an 'index key' for key access
Bugzilla: 1014573 985961
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Consolidate the concept of an 'index key' for accessing keys.  The index key
is the search term needed to find a key directly - basically the key type and
the key description.  We can add to that the description length.

This will be useful when turning a keyring into an associative array rather
than just a pointer block.

Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>