[crypto] x.509: Remove certificate date checks
Message-id: <20131002135405.13612.9838.stgit@warthog.procyon.org.uk>
Patchwork-id: 68312
O-Subject: [RHEL7 PATCH 22/33] X.509: Remove certificate date checks
Bugzilla: 985961
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Remove the certificate date checks that are performed when a certificate is
parsed.  There are two checks: a valid from and a valid to.  The first check is
causing a lot of problems with system clocks that don't keep good time and the
second places an implicit expiry date upon the kernel when used for module
signing, so do we really need them?

Signed-off-by: David Howells <dhowells@redhat.com>
cc: David Woodhouse <dwmw2@infradead.org>
cc: Rusty Russell <rusty@rustcorp.com.au>
cc: Josh Boyer <jwboyer@redhat.com>
cc: Alexander Holler <holler@ahsoftware.de>
cc: stable@vger.kernel.org
Signed-off-by: Jarod Wilson <jarod@redhat.com>